+++ changelog.md
@@ -1,3 +1,2 @@
 ### Additional Lanterns 1.0.2
-- Updated to core library 1.1
-- Improved the model transformations of the lanterns when held
+- Initial release of Additional Lanterns for Minecraft 1.19.3